tests/submission/Infraware/Forms/contents/Forms/option_constructor_005.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>option_constructor_005</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype">
		  <option value="1"> Miner </option>
		  <option value="2"> Puffer </option>
		  <option value="3"> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		var item = select.item(2);

		var newitem = new Option('CustomText', '6', true, true);
		select.add(newitem, item);

		if (select.item(2).text == "CustomText" && select.item(2).value == "6" && 
			select.item(2).defaultSelected == true && select.item(2).selected == true) {
			test(function() {
				assert_true(true, "calling of constructor method is successed.");
			});
		} else {
			test(function() {
				assert_unreached("calling of constructor method is failed.");
			});
		}

	</script>

 </body>
</html>
